requirements.txt
asn1crypto==0.24.0
bootstrap4==0.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-forms-bootstrap==3.1.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.10.1
enum34==1.1.6
idna==2.6
ipaddress==1.0.17
jeepney==0.4
keyring==10.6.0
keyrings.alt==3.0
mysqlclient==1.4.2.post1
oauthlib==3.0.2
Pillow==6.1.0
pycparser==2.19
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
pywin32-ctypes==0.2.0
pyxdg==0.25
requests==2.22.0
requests-oauthlib==1.2.0
SecretStorage==2.3.1
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
stripe==2.32.1
urllib3==1.25.3
